requirements.txt
flask==3.0.0
gtts==2.4.0
PyPDF2==3.0.1
python-docx==1.0.1
groq==0.4.2
gunicorn==21.2.0
